+++ project_cylon/Logger.py
@@ -1,10 +1,27 @@
 # -*- coding: utf-8 -*-
+import os
 import sys
+import shutil
 
 class Logger:
 
+    driver = None
+
+    screenshot_path = "./screenshots"
+    screenshot_number = 1
+    screenshot_enabled = False
+
     @classmethod
-    def tracebacklimit(cls, value=1000):
+    def enable_screenshots(cls):
+        cls.screenshot_enabled = True
+
+        if os.path.exists(cls.screenshot_path):
+            shutil.rmtree(cls.screenshot_path)
+
+        os.makedirs(cls.screenshot_path)
+
+    @classmethod
+    def tracebacklimit(cls, value=0):
         sys.tracebacklimit = int(value)
 
     @classmethod
@@ -18,11 +35,19 @@ class Logger:
         raise RuntimeError
 
     @classmethod
-    def failed(cls, message, actual="--", expect="--"):
+    def failed(cls, message, actual="", expect=""):
         message = "FAIL: %s\n" % message
 
-        if actual != "--" and expect != "--":
+        if actual != "" and expect != "":
             message += "actual: '%s'\nexpect: '%s'\n" % (actual, expect)
 
         print message
+
+        if cls.screenshot_enabled == True:
+            filename = "%s/%03d.png" % (cls.screenshot_path, cls.screenshot_number)
+            cls.driver.save_screenshot(filename)
+            cls.screenshot_number += 1
+
+            print "Save screenshot to '%s'" % filename
+
         raise RuntimeError