+++ project_cylon/behook.py
@@ -7,6 +7,7 @@ from behave import __main__
 configuration.parser.add_argument('--browser')
 configuration.parser.add_argument('--site')
 configuration.parser.add_argument('--debug', action='store_true')
+configuration.parser.add_argument('--screenshot', action='store_true')
 
 ## command that run behave.
 def main():